+++ packages/substrate-bindings/src/codecs/metadata/metadata.ts
@@ -1,6 +1,18 @@
-import { Enum, Struct, u32, Codec, createCodec } from "scale-ts"
-import { v15 } from "./v15"
+import {
+  Bytes,
+  type Codec,
+  type CodecType,
+  Enum,
+  Option,
+  Struct,
+  Tuple,
+  compact,
+  createCodec,
+  u32,
+} from "scale-ts"
+import type { HexString } from "../scale"
 import { v14 } from "./v14"
+import { v15 } from "./v15"
 
 const unsupportedFn = () => {
   throw new Error("Unsupported metadata version!")
@@ -32,3 +44,32 @@ export const metadata = Struct({
     v15,
   }),
 })
+
+const opaqueBytes = Bytes()
+const optionOpaque = Option(opaqueBytes)
+const opaqueOpaqueBytes = Tuple(compact, opaqueBytes)
+
+export const decAnyMetadata = (
+  input: Uint8Array | HexString,
+): CodecType<typeof metadata> => {
+  try {
+    return metadata.dec(input)
+  } catch (_) {}
+
+  // comes from metadata.metadata_at_version
+  try {
+    return metadata.dec(optionOpaque.dec(input)!)
+  } catch (_) {}
+
+  // comes from state.getMetadata
+  try {
+    return metadata.dec(opaqueBytes.dec(input))
+  } catch (_) {}
+
+  // comes from metadata.metadata
+  try {
+    return metadata.dec(opaqueOpaqueBytes.dec(input)[1])
+  } catch (_) {}
+
+  throw null
+}